+++ authenticator/manage-server.py
@@ -2,15 +2,25 @@
 
 """
 Simple script to manage virtual servers.
+
+This needs the same setup as mumble-authenticator.py.
 """
 
 import sys
 import Ice
 
+# Config
+
+ice_slice = '/usr/share/slice/Murmur.ice'
+ice_host = '127.0.0.1'
+ice_port = 6502
+
+
+# Run
 
 try:
     # noinspection PyArgumentList
-    Ice.loadSlice('', ['-I' + Ice.getSliceDir(), '/usr/share/slice/Murmur.ice'])
+    Ice.loadSlice('', ['-I' + Ice.getSliceDir(), ice_slice])
 except RuntimeError as e:
     print(format(e))
     sys.exit(0)
@@ -18,7 +28,7 @@ except RuntimeError as e:
 import Murmur
 
 ice = Ice.initialize(sys.argv)
-meta = Murmur.MetaPrx.checkedCast(ice.stringToProxy('Meta -e 1.0:tcp -h 127.0.0.1 -p 6502'))
+meta = Murmur.MetaPrx.checkedCast(ice.stringToProxy('Meta -e 1.0:tcp -h %s -p %s' % (ice_host, ice_port)))
 
 
 if len(sys.argv) > 1 and sys.argv[1] == 'list':
@@ -36,7 +46,11 @@ elif len(sys.argv) > 1 and sys.argv[1] == 'conf-all':
         print('usage: manage-server.py conf-all 2')
     else:
         server = meta.getServer(int(sys.argv[2]))
-        print(server.getAllConf())
+        if server is None:
+            print('invalid id')
+        else:
+            for key, value in server.getAllConf().items():
+                print('{0}: {1}'.format(key, value))
 
 elif len(sys.argv) > 1 and sys.argv[1] == 'create':
     new_server = meta.newServer()
@@ -47,8 +61,11 @@ elif len(sys.argv) > 1 and sys.argv[1] == 'pw':
         print('usage: manage-server.py pw 2 super-user-pw')
     else:
         server = meta.getServer(int(sys.argv[2]))
-        server.setSuperuserPassword(sys.argv[3])
-        print('ok')
+        if server is None:
+            print('invalid id')
+        else:
+            server.setSuperuserPassword(sys.argv[3])
+            print('done')
 
 elif len(sys.argv) > 1 and sys.argv[1] == 'set-conf':
     if len(sys.argv) < 5:
@@ -56,24 +73,33 @@ elif len(sys.argv) > 1 and sys.argv[1] == 'set-conf':
         print('valid names: port, registerName, welcometext, ...')
     else:
         server = meta.getServer(int(sys.argv[2]))
-        server.setConf(sys.argv[3], sys.argv[4])
-        print('ok')
+        if server is None:
+            print('invalid id')
+        else:
+            server.setConf(sys.argv[3], sys.argv[4])
+            print('done')
 
 elif len(sys.argv) > 1 and sys.argv[1] == 'start':
     if len(sys.argv) < 3:
         print('usage: manage-server.py start 2')
     else:
         server = meta.getServer(int(sys.argv[2]))
-        server.start()
-        print('ok')
+        if server is None:
+            print('invalid id')
+        else:
+            server.start()
+            print('done')
 
 elif len(sys.argv) > 1 and sys.argv[1] == 'stop':
     if len(sys.argv) < 3:
         print('usage: manage-server.py stop 2')
     else:
         server = meta.getServer(int(sys.argv[2]))
-        server.stop()
-        print('ok')
+        if server is None:
+            print('invalid id')
+        else:
+            server.stop()
+            print('done')
 
 elif len(sys.argv) > 1 and sys.argv[1] == 'delete':
     if len(sys.argv) < 3:
@@ -81,16 +107,18 @@ elif len(sys.argv) > 1 and sys.argv[1] == 'delete':
     else:
         server_id = int(sys.argv[2])
         if server_id == 1:
-            print('No, do not delete server 1.')
+            print('No, not deleting server 1.')
         else:
-            sys.stdout.write('Are you sure? [yes/No] ')
-            choice = input()
-            if choice == 'yes':
-                server = meta.getServer(server_id)
-                server.delete()
-                print('deleted server')
+            server = meta.getServer(server_id)
+            if server is None:
+                print('invalid id')
             else:
-                print('no')
+                sys.stdout.write('Are you sure? [yes/No] ')
+                if input().lower() == 'yes':
+                    server.delete()
+                    print('deleted server')
+                else:
+                    print('no')
 
 else:
     print('usage: manage-server.py list|conf-all|create|pw|set-conf|start|stop|delete')